# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AVIS — Agentic Visual Incident Investigation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Video {
  id: number;
  fileName: string;
  cameraName: string;
  status: string;
  uploadTime: string;
  /** @nullable */
  processedAt?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  frameCount?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export interface VideoInput {
  fileName: string;
  cameraName: string;
  durationSeconds?: number;
}

export interface Detection {
  id: number;
  videoId: number;
  objectType: string;
  confidence: number;
  frameNumber: number;
  timestamp: string;
  /** @nullable */
  boundingBox?: string | null;
  createdAt: string;
}

export interface DetectionInput {
  videoId: number;
  objectType: string;
  confidence: number;
  frameNumber: number;
  timestamp: string;
  boundingBox?: string;
}

export interface TrackingRecord {
  id: number;
  videoId: number;
  personId: string;
  /** @nullable */
  path?: string | null;
  camera: string;
  /** @nullable */
  durationSeconds?: number | null;
  createdAt: string;
}

export interface TrackingInput {
  videoId: number;
  personId: string;
  path?: string;
  camera: string;
  durationSeconds?: number;
}

export interface IncidentEvent {
  id: number;
  videoId: number;
  eventType: string;
  riskScore: number;
  severity: string;
  description: string;
  timestamp: string;
  camera: string;
  /** @nullable */
  frameNumber?: number | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export interface EventInput {
  videoId: number;
  eventType: string;
  riskScore: number;
  severity: string;
  description: string;
  timestamp: string;
  camera: string;
  frameNumber?: number;
  metadata?: string;
}

export interface Report {
  id: number;
  /** @nullable */
  videoId?: number | null;
  title: string;
  summary: string;
  findings: string;
  recommendations: string;
  severity: string;
  riskScore: number;
  /** @nullable */
  timeline?: string | null;
  createdAt: string;
}

export interface ReportInput {
  videoId?: number;
  title: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface DashboardStats {
  totalVideos: number;
  totalDetections: number;
  activeIncidents: number;
  criticalIncidents: number;
  totalReports: number;
  processedVideos: number;
}

export interface TimelineEntry {
  id: number;
  timestamp: string;
  eventType: string;
  description: string;
  severity: string;
  camera: string;
  riskScore: number;
}

export interface RiskCount {
  severity: string;
  count: number;
}

export interface Settings {
  loiteringThresholdMinutes: number;
  crowdThreshold: number;
  runningVelocityThreshold: number;
  abandonedObjectMinutes: number;
  defaultModel: string;
  enableMotionDetection: boolean;
  enableFaceDetection: boolean;
  riskScoreLoitering: number;
  riskScoreRestrictedArea: number;
  riskScoreAbandonedBag: number;
  riskScoreCrowd: number;
  riskScoreRunning: number;
}

export interface SettingsUpdate {
  loiteringThresholdMinutes?: number;
  crowdThreshold?: number;
  runningVelocityThreshold?: number;
  abandonedObjectMinutes?: number;
  defaultModel?: string;
  enableMotionDetection?: boolean;
  enableFaceDetection?: boolean;
  riskScoreLoitering?: number;
  riskScoreRestrictedArea?: number;
  riskScoreAbandonedBag?: number;
  riskScoreCrowd?: number;
  riskScoreRunning?: number;
}

export type ListDetectionsParams = {
videoId?: number;
objectType?: string;
limit?: number;
};

export type ListTrackingParams = {
videoId?: number;
personId?: string;
};

export type ListEventsParams = {
videoId?: number;
severity?: string;
eventType?: string;
limit?: number;
};