# lib/api-zod/src/generated/types/detectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AVIS — Agentic Visual Incident Investigation System API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionInput {
  videoId: number;
  objectType: string;
  confidence: number;
  frameNumber: number;
  timestamp: string;
  boundingBox?: string;
}